+++ QuebecTaxCalculator/src/main/java/dev/paintilya/taxcalculator/BLL/Control/QuebecTaxController.java
@@ -13,11 +13,11 @@ public class QuebecTaxController {
         taxBrackets.add(new TaxBracket(0, 46295, 0.15f));
         taxBrackets.add(new TaxBracket(46295, 92580, 0.20f));
         taxBrackets.add(new TaxBracket(92580, 112655, 0.24f));
-        taxBrackets.add(new TaxBracket(112655, Float.MAX_VALUE, 0.2575f));
+        taxBrackets.add(new TaxBracket(112655, Double.MAX_VALUE, 0.2575f));
     }
 
-    public float calculateTax(float netIncome) {
-        float tax = 0;
+    public double calculateTax(double netIncome) {
+        double tax = 0;
         for (TaxBracket bracket : taxBrackets) {
             if (netIncome > bracket.getMax()) {
                 tax += (bracket.getMax() - bracket.getMin()) * bracket.getRate();
