+++ QuebecTaxCalculator/src/main/java/dev/paintilya/taxcalculator/FEL/AppDriver.java
@@ -5,8 +5,8 @@ import dev.paintilya.taxcalculator.BLL.Control.QuebecTaxController;
 public class AppDriver {
     public static void main(String[] args) {
         QuebecTaxController taxController = new QuebecTaxController();
-        float netIncome = 50000;
-        float tax = taxController.calculateTax(netIncome);
+        double netIncome = 50000;
+        double tax = taxController.calculateTax(netIncome);
         String formattedTax = String.format("%.2f", tax);
         System.out.println("Pour un revenu net de " + netIncome + "$, la taxe à payer est de " + formattedTax + "$.");
     }
